Move upload files logic to ProjectView.
