+++ frontend/src/views/ProjectView.vue
@@ -1,11 +1,63 @@
 <template>
-  Lorem ipsum dolor sit amet, officia excepteur ex fugiat reprehenderit enim labore culpa sint ad
-  nisi Lorem pariatur mollit ex esse exercitation amet. Nisi anim cupidatat excepteur officia.
-  Reprehenderit nostrud nostrud ipsum Lorem est aliquip amet voluptate voluptate dolor minim nulla
-  est proident. Nostrud officia pariatur ut officia. Sit irure elit esse ea nulla sunt ex occaecat
-  reprehenderit commodo officia dolor Lorem duis laboris cupidatat officia voluptate. Culpa proident
-  adipisicing id nulla nisi laboris ex in Lorem sunt duis officia eiusmod. Aliqua reprehenderit
-  commodo ex non excepteur duis sunt velit enim. Voluptate laboris sint cupidatat ullamco ut ea
-  consectetur et est culpa et culpa duis.
-  <template> </template>
+  <div v-if="hasErorr" class="alert alert-danger alert-dismissible" role="alert">
+    <div>{{ errorMessage }}</div>
+    <button type="button" @click="dismissAlert" class="btn-close" aria-label="Close"></button>
+  </div>
+  <DropToStartField @dropFilesHandler="dropFilesHandler" />
 </template>
+
+<script setup lang="ts">
+import DropToStartField from '@/components/DropToStartField.vue'
+import type { User, Project } from '@/constants'
+import { Constants } from '@/constants'
+import axios from 'axios'
+import { ref } from 'vue'
+
+const props = defineProps<{ project: Project | undefined }>()
+const project = ref(props.project)
+const hasErorr = ref(false)
+const errorMessage = ref('')
+function dismissAlert() {
+  hasErorr.value = false
+}
+
+async function createUserAndProject() {
+  const env = await Constants.env
+  await axios.post<User>(`${env.NAUTILUS_WEBAPI}/users`)
+  const projectRequestData = {
+    name: 'First Project'
+  }
+  const createProjectResponse = await axios.post<Project>(
+    `${env.NAUTILUS_WEBAPI}/projects`,
+    projectRequestData
+  )
+  return createProjectResponse.data
+}
+
+async function dropFilesHandler(event: DragEvent) {
+  if (event.dataTransfer?.files == undefined) {
+    return
+  }
+  const files = event.dataTransfer.files
+  const env = await Constants.env
+  try {
+    if (project.value == undefined) {
+      project.value = await createUserAndProject()
+    }
+    const uploadFileRequestsList = []
+    for (let i = 0; i < files.length; ++i) {
+      const file = files[i]
+      const requestData = new FormData()
+      requestData.append('project_id', project.value.id)
+      requestData.append('uploaded_file', file)
+      uploadFileRequestsList.push(
+        axios.post(`${env.NAUTILUS_WEBAPI}/projects/${project.value.id}/files`, requestData)
+      )
+    }
+    await axios.all(uploadFileRequestsList)
+  } catch (error: any) {
+    hasErorr.value = true
+    errorMessage.value = `${error.request.responseURL}: ${error.response.statusText}`
+  }
+}
+</script>
